+++ src/transactions/test/test_helper/ReviewPreIssuanceRequestHelper.h
@@ -30,6 +30,7 @@ namespace txtest
     };
 	class ReviewPreIssuanceRequestHelper : public ReviewRequestHelper
 	{
+		uint32_t mTasksToRemove = 0;
 	public:
 		ReviewPreIssuanceRequestHelper(TestManager::pointer testManager);
 
@@ -39,6 +40,16 @@ namespace txtest
 
 		ReviewRequestResult applyReviewRequestTx(Account & source, uint64_t requestID, ReviewRequestOpAction action, std::string rejectReason,
 			ReviewRequestResultCode expectedResult = ReviewRequestResultCode::SUCCESS);
+
+		ReviewRequestResult
+		applyReviewRequestTxWithTasks(Account & source, uint64_t requestID, ReviewRequestOpAction action,
+							 std::string rejectReason, uint32_t tasksToRemove, ReviewRequestResultCode expectedResult =
+							 		ReviewRequestResultCode::SUCCESS);
+
+		TransactionFramePtr
+		createReviewRequestTx(Account &source, uint64_t requestID, Hash requestHash,
+							  ReviewableRequestType requestType, ReviewRequestOpAction action,
+							  std::string rejectReason) override;
 	};
 }
 }