+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Check for exceeding of maxIssuanceAmount on pre issuance (add pendingIssuance of asset to sum in canAddAvailableForIssuance method)
 * Check for exceeding of maxIssuanceAmount on create `CREATE_SALE` request
 * Remove checking for not empty signers on signature validation
+* Return error code on review pre issuance if max issuance amount exceeded
 
 ## 3.0.0-x.7